pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linuxcnc-grpc-server"
version = "0.5.0"
description = "gRPC server exposing LinuxCNC machine control and HAL functionality"
readme = "README.md"
license = {text = "GPL-2.0-or-later"}
requires-python = ">=3.8"
keywords = ["linuxcnc", "cnc", "grpc", "machine-control", "hal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = ["grpcio-tools>=1.60.0", "pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
linuxcnc-grpc-server = "linuxcnc_grpc_server.server:main"

[project.urls]
Repository = "https://github.com/dougcalobrisi/linuxcnc-grpc-server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"linuxcnc_grpc_server.proto" = ["*.proto"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/linuxcnc_grpc_server"]
omit = ["*/_generated/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
